app/requirements.txt
boto3==1.22.1
Flask==2.0.2
mysql-connector-python==8.0.28
mysql_connector_repackaged==0.3.1
python-dotenv==0.20.0
